Assignment 3 Initial commit
Initial commit, stripped out most of the junk from previous assignments to pretty up the layout. Slapped the thing in some divs so it follows more of an itch.io format. Centered the main contents so that it is more appealing to me. Changed the font to something actually decent.